+++ lib/pages/journal/infinite_journal_page.dart
@@ -1,55 +1,45 @@
 import 'dart:async';
 
-import 'package:flutter/cupertino.dart';
 import 'package:flutter/material.dart';
-import 'package:flutter/services.dart';
-import 'package:flutter_gen/gen_l10n/app_localizations.dart';
+import 'package:flutter_bloc/flutter_bloc.dart';
 import 'package:infinite_scroll_pagination/infinite_scroll_pagination.dart';
+import 'package:lotti/blocs/journal/journal_page_cubit.dart';
 import 'package:lotti/classes/journal_entities.dart';
 import 'package:lotti/classes/tag_type_definitions.dart';
 import 'package:lotti/database/database.dart';
 import 'package:lotti/get_it.dart';
 import 'package:lotti/themes/theme.dart';
 import 'package:lotti/utils/platform.dart';
+import 'package:lotti/widgets/app_bar/journal_sliver_appbar.dart';
 import 'package:lotti/widgets/create/add_actions.dart';
 import 'package:lotti/widgets/journal/journal_card.dart';
-import 'package:lotti/widgets/journal/tags/selected_tags_widget.dart';
-import 'package:lotti/widgets/misc/search_widget.dart';
-import 'package:multi_select_flutter/util/multi_select_item.dart';
 
-class FilterBy {
-  FilterBy({
-    required this.typeName,
-    required this.name,
+class JournalPageWrapper extends StatelessWidget {
+  const JournalPageWrapper({
+    super.key,
+    this.navigatorKey,
   });
 
-  final String typeName;
-  final String name;
+  final GlobalKey? navigatorKey;
+
+  @override
+  Widget build(BuildContext context) {
+    return Navigator(
+      key: navigatorKey,
+      onGenerateRoute: (RouteSettings settings) {
+        return MaterialPageRoute(
+          settings: settings,
+          builder: (BuildContext context) {
+            return BlocProvider<JournalPageCubit>(
+              create: (BuildContext context) => JournalPageCubit(),
+              child: const InfiniteJournalPage(),
+            );
+          },
+        );
+      },
+    );
+  }
 }
-
-final List<String> defaultTypes = [
-  'JournalEntry',
-  'JournalAudio',
-  'JournalImage',
-  'SurveyEntry',
-  'Task',
-  // 'QuantitativeEntry',
-  'MeasurementEntry',
-  'WorkoutEntry',
-  // 'HabitCompletionEntry',
-];
-
-final List<FilterBy> _entryTypes = [
-  FilterBy(typeName: 'Task', name: 'Task'),
-  FilterBy(typeName: 'JournalEntry', name: 'Text'),
-  FilterBy(typeName: 'JournalAudio', name: 'Audio'),
-  FilterBy(typeName: 'JournalImage', name: 'Photo'),
-  FilterBy(typeName: 'MeasurementEntry', name: 'Measured'),
-  FilterBy(typeName: 'SurveyEntry', name: 'Survey'),
-  FilterBy(typeName: 'WorkoutEntry', name: 'Workout'),
-  FilterBy(typeName: 'HabitCompletionEntry', name: 'Habit'),
-  FilterBy(typeName: 'QuantitativeEntry', name: 'Quant'),
-];
 
 class InfiniteJournalPage extends StatefulWidget {
   const InfiniteJournalPage({
@@ -66,21 +56,10 @@ class InfiniteJournalPage extends StatefulWidget {
 class _InfiniteJournalPageState extends State<InfiniteJournalPage> {
   final JournalDb _db = getIt<JournalDb>();
 
-  late Set<String> types;
-
   StreamController<List<TagEntity>> matchingTagsController =
       StreamController<List<TagEntity>>();
 
-  final List<MultiSelectItem<FilterBy?>> _items = _entryTypes
-      .map((entryType) => MultiSelectItem<FilterBy?>(entryType, entryType.name))
-      .toList();
-
   Set<String> tagIds = {};
-  bool starredEntriesOnly = false;
-  bool flaggedEntriesOnly = false;
-  bool privateEntriesOnly = false;
-  bool showPrivateEntriesSwitch = false;
-
   static const _pageSize = 50;
 
   final PagingController<int, JournalEntity> _pagingController =
@@ -89,13 +68,12 @@ class _InfiniteJournalPageState extends State<InfiniteJournalPage> {
   @override
   void initState() {
     _pagingController.addPageRequestListener(_fetchPage);
-
-    types = defaultTypes.toSet();
-
     super.initState();
   }
 
   Future<void> _fetchPage(int pageKey) async {
+    final cubit = context.read<JournalPageCubit>();
+
     try {
       Set<String>? entryIds;
       for (final tagId in tagIds) {
@@ -107,19 +85,24 @@ class _InfiniteJournalPageState extends State<InfiniteJournalPage> {
         }
       }
 
+      final types = cubit.state.selectedEntryTypes
+          .map((e) => e?.typeName)
+          .whereType<String>()
+          .toList();
+
       final newItems = await _db
           .watchJournalEntities(
-            types: types.toList(),
+            types: types,
             ids: entryIds?.toList(),
-            starredStatuses: starredEntriesOnly ? [true] : [true, false],
-            privateStatuses: privateEntriesOnly ? [true] : [true, false],
-            flaggedStatuses: flaggedEntriesOnly ? [1] : [1, 0],
+            starredStatuses:
+                cubit.state.starredEntriesOnly ? [true] : [true, false],
+            privateStatuses:
+                cubit.state.privateEntriesOnly ? [true] : [true, false],
+            flaggedStatuses: cubit.state.flaggedEntriesOnly ? [1] : [1, 0],
             limit: _pageSize,
             offset: pageKey,
           )
           .first;
-
-      //final newItems = await RemoteApi.getCharacterList(pageKey, _pageSize);
 
       final isLastPage = newItems.length < _pageSize;
       if (isLastPage) {
@@ -151,185 +134,47 @@ class _InfiniteJournalPageState extends State<InfiniteJournalPage> {
     });
   }
 
-  Widget searchRow() {
-    final localizations = AppLocalizations.of(context)!;
-
-    return Column(
-      mainAxisSize: MainAxisSize.min,
-      children: [
-        Column(
-          children: [
-            Wrap(
-              spacing: 4,
-              runSpacing: 4,
-              children: [
-                ..._items.map(
-                  (MultiSelectItem<FilterBy?> item) => GestureDetector(
-                    onTap: () {
-                      setState(() {
-                        final typeName = item.value?.typeName;
-                        if (typeName != null) {
-                          if (types.contains(typeName)) {
-                            types.remove(typeName);
-                          } else {
-                            types.add(typeName);
-                          }
-                          resetQuery();
-                          HapticFeedback.heavyImpact();
-                        }
-                      });
-                    },
-                    child: MouseRegion(
-                      cursor: SystemMouseCursors.click,
-                      child: ClipRRect(
-                        borderRadius: BorderRadius.circular(4),
-                        child: ColoredBox(
-                          color: types.contains(item.value?.typeName)
-                              ? styleConfig().selectedChoiceChipColor
-                              : styleConfig().unselectedChoiceChipColor,
-                          child: Padding(
-                            padding: const EdgeInsets.symmetric(
-                              vertical: 1,
-                              horizontal: 8,
-                            ),
-                            child: Text(
-                              item.label,
-                              style: TextStyle(
-                                fontFamily: 'Oswald',
-                                fontSize: 14,
-                                color: types.contains(item.value?.typeName)
-                                    ? styleConfig().selectedChoiceChipTextColor
-                                    : styleConfig()
-                                        .unselectedChoiceChipTextColor,
-                              ),
-                            ),
-                          ),
-                        ),
-                      ),
-                    ),
-                  ),
-                ),
-              ],
-            ),
-            Padding(
-              padding: const EdgeInsets.symmetric(horizontal: 8),
-              child: Row(
-                mainAxisAlignment: MainAxisAlignment.spaceAround,
-                children: [
-                  Visibility(
-                    visible: showPrivateEntriesSwitch,
-                    child: Row(
-                      children: [
-                        Text(
-                          localizations.journalPrivateTooltip,
-                          style: TextStyle(
-                            color: styleConfig().primaryTextColor,
-                          ),
-                        ),
-                        CupertinoSwitch(
-                          value: privateEntriesOnly,
-                          activeColor: styleConfig().private,
-                          onChanged: (bool value) {
-                            setState(() {
-                              privateEntriesOnly = value;
-                              resetQuery();
-                            });
-                          },
-                        ),
-                      ],
-                    ),
-                  ),
-                  Text(
-                    localizations.journalFavoriteTooltip,
-                    style: TextStyle(color: styleConfig().primaryTextColor),
-                  ),
-                  CupertinoSwitch(
-                    value: starredEntriesOnly,
-                    activeColor: styleConfig().starredGold,
-                    onChanged: (bool value) {
-                      setState(() {
-                        starredEntriesOnly = value;
-                        resetQuery();
-                      });
-                    },
-                  ),
-                  Text(
-                    localizations.journalFlaggedTooltip,
-                    style: TextStyle(color: styleConfig().primaryTextColor),
-                  ),
-                  CupertinoSwitch(
-                    value: flaggedEntriesOnly,
-                    activeColor: styleConfig().starredGold,
-                    onChanged: (bool value) {
-                      setState(() {
-                        flaggedEntriesOnly = value;
-                        resetQuery();
-                      });
-                    },
-                  ),
-                ],
-              ),
-            ),
-            SelectedTagsWidget(
-              removeTag: removeTag,
-              tagIds: tagIds.toList(),
-            ),
-          ],
-        )
-      ],
-    );
-  }
-
   @override
   Widget build(BuildContext context) {
-    return Navigator(
-      key: widget.navigatorKey,
-      onGenerateRoute: (RouteSettings settings) {
-        return MaterialPageRoute(
-          settings: settings,
-          builder: (BuildContext context) {
-            return Scaffold(
-              backgroundColor: styleConfig().negspace,
-              appBar: JournalPageAppBar(
-                title: 'Journal',
-                match: '',
-                onQueryChanged: (text) {},
-                searchRow: searchRow(),
-              ),
-              body: RefreshIndicator(
-                onRefresh: () => Future.sync(_pagingController.refresh),
-                child: PagedListView<int, JournalEntity>(
-                  pagingController: _pagingController,
-                  builderDelegate: PagedChildBuilderDelegate<JournalEntity>(
-                    itemBuilder: (context, item, index) {
-                      return item.maybeMap(
-                        journalImage: (JournalImage image) {
-                          return JournalImageCard(
-                            item: image,
-                            key: ValueKey(item.meta.id),
-                          );
-                        },
-                        orElse: () {
-                          return JournalCard(
-                            item: item,
-                            key: ValueKey(item.meta.id),
-                          );
-                        },
+    return Scaffold(
+      backgroundColor: styleConfig().negspace,
+      floatingActionButton: RadialAddActionButtons(
+        radius: isMobile ? 180 : 120,
+        isMacOS: isMacOS,
+        isIOS: isIOS,
+        isAndroid: isAndroid,
+      ),
+      body: RefreshIndicator(
+        onRefresh: () => Future.sync(_pagingController.refresh),
+        child: CustomScrollView(
+          slivers: <Widget>[
+            JournalSliverAppBar(
+              resetQuery: resetQuery,
+            ),
+            PagedSliverList<int, JournalEntity>(
+              pagingController: _pagingController,
+              builderDelegate: PagedChildBuilderDelegate<JournalEntity>(
+                itemBuilder: (context, item, index) {
+                  return item.maybeMap(
+                    journalImage: (JournalImage image) {
+                      return JournalImageCard(
+                        item: image,
+                        key: ValueKey(item.meta.id),
                       );
                     },
-                  ),
-                ),
+                    orElse: () {
+                      return JournalCard(
+                        item: item,
+                        key: ValueKey(item.meta.id),
+                      );
+                    },
+                  );
+                },
               ),
-              floatingActionButton: RadialAddActionButtons(
-                radius: isMobile ? 180 : 120,
-                isMacOS: isMacOS,
-                isIOS: isIOS,
-                isAndroid: isAndroid,
-              ),
-            );
-          },
-        );
-      },
+            ),
+          ],
+        ),
+      ),
     );
   }
 
@@ -337,39 +182,5 @@ class _InfiniteJournalPageState extends State<InfiniteJournalPage> {
   void dispose() {
     _pagingController.dispose();
     super.dispose();
-  }
-}
-
-class JournalPageAppBar extends StatelessWidget with PreferredSizeWidget {
-  const JournalPageAppBar({
-    super.key,
-    required this.title,
-    required this.match,
-    required this.onQueryChanged,
-    required this.searchRow,
-  });
-
-  final String title;
-  final String match;
-  final void Function(String) onQueryChanged;
-  final Widget searchRow;
-
-  @override
-  Size get preferredSize => const Size.fromHeight(kToolbarHeight * 3);
-
-  @override
-  Widget build(BuildContext context) {
-    return Column(
-      mainAxisAlignment: MainAxisAlignment.spaceBetween,
-      children: [
-        const SizedBox(height: 5),
-        SearchWidget(
-          text: match,
-          onChanged: onQueryChanged,
-          hintText: 'Search $title...',
-        ),
-        searchRow,
-      ],
-    );
   }
 }