+++ CHANGELOG.md
@@ -12,6 +12,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Hide task title label when task defined
 - Floating search bar replaced in settings
 - Floating search bar replaced in journal
+- New search header in journal
 
 ## [0.8.225] - 2023-01-06
 ### Changed: